scripts/requirements.txt
PyYAML>=6.0
pydantic>=2.0
Jinja2>=3.1
